+++ dms/models/__init__.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 ###################################################################################
 # 
 #    Copyright (C) 2017 MuK IT GmbH
@@ -19,10 +17,15 @@
 #
 ###################################################################################
 
-from . import dms_base
+from . import dms_model
+from . import dms_locking
+from . import dms_access
 from . import dms_settings
 from . import dms_directory
 from . import dms_file
 from . import dms_data
-from . import dms_lock
-from . import res_config
+from . import dms_tag
+from . import dms_category
+from . import dms_data_database
+from . import ir_http
+from . import res_config_settings