+++ dms/tests/dms_case.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 ###################################################################################
 # 
 #    Copyright (C) 2017 MuK IT GmbH
@@ -22,9 +20,8 @@
 import os
 import base64
 import logging
-import unittest
 
-from odoo import _
+from odoo import SUPERUSER_ID
 from odoo.tests import common
 
 _path = os.path.dirname(os.path.dirname(__file__))
@@ -32,17 +29,19 @@ _logger = logging.getLogger(__name__)
 
 class DMSTestCase(common.TransactionCase):
     
-    at_install = True
-    post_install = False
+    at_install = False
+    post_install = True
     
     def setUp(self):
         super(DMSTestCase, self).setUp()
-        self.demouser = self.browse_ref("base.user_demo")
-        self.dmsuser = self.browse_ref("muk_dms.user_dmsuser_demo")
-        self.dmsmanager = self.browse_ref("muk_dms.user_dmsmanager_demo")
+        self.adminuser = SUPERUSER_ID
+        self.demouser = self.browse_ref("base.user_demo").id
+        self.settings = self.env['muk_dms.settings']
+        self.directory = self.env['muk_dms.directory']
+        self.file = self.env['muk_dms.file']
         
     def tearDown(self):
         super(DMSTestCase, self).tearDown()
         
     def file_base64(self):
-        return base64.b64encode(b"Hello World!")
+        return base64.b64encode(b"\xff data")