+++ dms/doc/changelog.rst
@@ -1,8 +1,20 @@
+`2.1.0`
+-------
+
+- Added small and medium thumbnail size fields
+
+`2.0.0`
+-------
+
+- Migrated to Python 3
+- Updated dependencies
+- Updated document views and widgets
+- Updated database models (no automatic migration!)
+
 `1.3.0`
 -------
 
 - Updated dependencies
-
 
 `1.2.0`
 -------